+++ yudao-module-system/yudao-module-system-impl/src/main/java/cn/iocoder/yudao/module/system/dal/mysql/sms/SmsTemplateMapper.java
@@ -48,7 +48,7 @@ public interface SmsTemplateMapper extends BaseMapperX<SmsTemplateDO> {
                 .orderByDesc(SmsTemplateDO::getId));
     }
 
-    default Integer selectCountByChannelId(Long channelId) {
+    default Long selectCountByChannelId(Long channelId) {
         return selectCount(SmsTemplateDO::getChannelId, channelId);
     }
