+++ yudao-module-system/yudao-module-system-impl/src/main/java/cn/iocoder/yudao/module/system/service/dict/DictDataServiceImpl.java
@@ -209,7 +209,7 @@ public class DictDataServiceImpl implements DictDataService {
     }
 
     @Override
-    public int countByDictType(String dictType) {
+    public long countByDictType(String dictType) {
         return dictDataMapper.selectCountByDictType(dictType);
     }
